# Conflicts:
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/presentation/fragment/AddEditProductPreviewFragment.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/presentation/viewmodel/AddEditProductPreviewViewModel.kt
#	features/merchant/product_add_edit/src/test/java/com/tokopedia/product/addedit/preview/presentation/viewmodel/AddEditProductPreviewViewModelTest.kt
#	features/merchant/product_add_edit/src/test/java/com/tokopedia/product/addedit/preview/presentation/viewmodel/AddEditProductPreviewViewModelTestFixture.kt
#	features/merchant/product_manage_common/src/main/java/com/tokopedia/product/manage/common/feature/variant/data/model/GetProductV3.kt
#	features/merchant/product_manage_common/src/main/java/com/tokopedia/product/manage/common/feature/variant/presentation/data/GetVariantResult.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/adapter/viewholder/ProductViewHolder.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/fragment/ProductManageFragment.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/mapper/ProductMapper.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/model/ViewState.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/ui/bottomsheet/ProductManageBottomSheet.kt
#	features/merchant/product_manage_list/src/main/java/com/tokopedia/product/manage/feature/list/view/viewmodel/ProductManageViewModel.kt
#	features/merchant/product_manage_list/src/main/res/values/strings.xml
#	features/merchant/seller_home/src/main/java/com/tokopedia/sellerhome/view/activity/SellerHomeActivity.kt
#	features/merchant/shop_common/src/main/java/com/tokopedia/shop/common/domain/interactor/model/adminrevamp/AdminInfoResponse.kt
#	features/transaction/sellerorder/src/main/java/com/tokopedia/sellerorder/list/presentation/viewmodels/SomListViewModel.kt
#	features/transaction/sellerorder/src/test/java/com/tokopedia/sellerorder/list/presentation/viewmodels/SomListViewModelTest.kt

+++ features/transaction/atc_common/src/main/java/com/tokopedia/atc_common/data/model/response/atcexternal/AddToCartExternalResponse.kt
@@ -25,7 +25,7 @@ data class DataResponse(
 
 data class AddToCartExternalDataResponse(
         @SerializedName("product_id")
-        val productId: Int = 0,
+        val productId: Long = 0,
         @SerializedName("product_name")
         val productName: String = "",
         @SerializedName("quantity")
@@ -45,7 +45,7 @@ data class AddToCartExternalDataResponse(
         @SerializedName("url")
         val url: String = "",
         @SerializedName("cart_id")
-        val cartId: Int = 0,
+        val cartId: Long = 0,
         @SerializedName("brand")
         val brand: String = "",
         @SerializedName("category_id")
